Adding comments y States patterns and try catch blocks in both threads

+++ src/clientServer/Worm.h
@@ -13,13 +13,41 @@
 
 namespace Worm {
 class Worm {
+    /**
+     * Fundamental class of the game, it is in charge of handling the user's
+     * entries, and delegate in their attributes the rendering and animation
+     */
    public:
     explicit Worm(SDL_Renderer &renderer);
     ~Worm() = default;
-    void update(double dt);
+    /**
+     * Calls State::update to change frame of animation
+     * @param dt
+     */
+    void update(float dt);
+    /**renderizar el gusano
+     * Render worm in position (x,y)
+     * @param x
+     * @param y
+     */
     void render(int x, int y);
+    /**
+     * Using a state pattern, change its state depending on the input, and
+     * sends it to the server
+     * @param key
+     * @param out
+     */
     void handleKeyDown(SDL_Keycode key, IO::Stream<IO::PlayerInput> *out);
+    /**
+     * Same as handleKeyDown, but stops its current status.
+     * @param key
+     * @param out
+     */
     void handleKeyUp(SDL_Keycode key, IO::Stream<IO::PlayerInput> *out);
+    /**
+     * Attributte that implements state pattern to change the behavior
+     * of the class polymorphically.
+     */
     std::shared_ptr<State> state;
 
    private:
