+++ src/clientServer/WormWalkRight.h
@@ -16,7 +16,7 @@ class WalkRight : public State {
     explicit WalkRight(SDL_Renderer &r);
     ~WalkRight();
     void render(int x, int y) override;
-    void update(double dt) override;
+    void update(float dt) override;
     IO::PlayerInput moveRight(Worm &w) override;
     IO::PlayerInput moveLeft(Worm &w) override;
     IO::PlayerInput stopMove(Worm &w) override;
